IsPassed has been modified to handle 32 bit numbers. Implementation of 5 baud init in KLineHAL.c has been started. Getting the input of a pin has been implemented in Ports.c

+++ src/Utilities/Utilities.h
@@ -9,8 +9,9 @@
 #define	UTILITIES_H
 
 #include "../types.h"
+#include "mainUtils.h"
 
-extern uint8 IsPassed(uint8 Time, uint8 limit);
+extern uint32 IsPassed(uint32 Time, uint32 limit);
 
 #endif	/* UTILITIES_H */
